refactor: put httparse into stacks-common so it can be used in other projects (like the signer binaries)
