+++ stacks-common/src/deps_common/mod.rs
@@ -16,3 +16,7 @@
 
 pub mod bech32;
 pub mod bitcoin;
+pub mod httparse;
+
+#[cfg(test)]
+pub mod httparse_tests;
