+++ stackslib/src/deps/mod.rs
@@ -15,7 +15,3 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pub mod ctrlc;
-pub mod httparse;
-
-#[cfg(test)]
-pub mod httparse_tests;
